pyproject.toml
[project]
name = "llm-redteam-suite"
version = "0.1.0"
description = "Adversarial prompt evaluation harness for LLM providers."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "httpx",
    "pyyaml",
    "python-dotenv",
    "rich",
    "pandas",
    "openai>=1.0",
    "huggingface_hub>=0.22",
]

[project.optional-dependencies]
dev = [
    "pytest",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["targets", "judge", "scripts"]
